UpdateScore added
